Update serial regression tests to run in parallel

+++ actions/clean-up/index.js
@@ -2,68 +2,55 @@ import { getExecOutput, exec } from '@actions/exec'
 import { getInput } from '@actions/core'
 const tests = [
   {
-    name: "type=embedded cluster, env=airgapped, phase=upgraded install, rbac=cluster admin",
     backend_config: "embedded-airgapped-upgrade-backend-config.tfvars",
     terraform_script: "embedded-airgapped-upgrade.sh",
   },
   {
-    name: "type=embedded cluster, env=airgapped, phase=new install, rbac=cluster admin",
     backend_config: "embedded-airgapped-install-backend-config.tfvars",
     terraform_script: "embedded-airgapped-install.sh"
   },
   {
-    name: "type=embedded cluster, env=airgapped, phase=new install, rbac=cluster admin",
     backend_config: "embedded-airgapped-install-backend-config-k3s.tfvars",
     terraform_script: "embedded-airgapped-install-k3s.sh",
     kubernetes_distro: "k3s"
   },
   {
-    name: "type=embedded cluster, env=online, phase=new install, rbac=cluster admin",
     backend_config: "embedded-online-install-backend-config.tfvars",
     terraform_script: "embedded-online-install.sh"
   },
   {
-    name: "type=embedded cluster, env=online, phase=upgraded install, rbac=cluster admin",
     backend_config: "embedded-online-upgrade-backend-config.tfvars",
     terraform_script: "embedded-online-upgrade.sh",
   },
   {
-    name: "type=existing cluster, env=airgapped, phase=new install, rbac=cluster admin",
     backend_config: "existing-airgapped-install-admin-backend-config.tfvars",
     terraform_script: "existing-airgapped-install-admin.sh"
   },
   {
-    name: "type=existing cluster, env=airgapped, phase=new install, rbac=minimal rbac",
     backend_config: "existing-airgapped-install-minimum-backend-config.tfvars",
     terraform_script: "existing-airgapped-install-minimum.sh"
   },
   {
-    name: "type=existing cluster, env=online, phase=upgraded install, rbac=cluster admin",
     backend_config: "existing-online-upgrade-admin-backend-config.tfvars",
     terraform_script: "existing-online-upgrade-admin.sh",
   },
   {
-    name: "type=existing cluster, env=online, phase=upgraded install, rbac=minimal rbac",
     backend_config: "existing-online-upgrade-minimum-backend-config.tfvars",
     terraform_script: "existing-online-upgrade-minimum.sh",
   },
   {
-    name: "type=existing cluster, env=online, phase=new install, rbac=cluster admin",
     backend_config: "existing-online-install-admin-backend-config.tfvars",
     terraform_script: "existing-online-install-admin.sh"
   },
   {
-    name: "type=existing cluster, env=online, phase=new install, rbac=minimal rbac",
     backend_config: "existing-online-install-minimum-backend-config.tfvars",
     terraform_script: "existing-online-install-minimum.sh"
   },
   {
-    name: "type=existing cluster, env=airgapped, phase=upgraded install, rbac=cluster admin",
     backend_config: "existing-airgapped-upgrade-admin-backend-config.tfvars",
     terraform_script: "existing-airgapped-upgrade-admin.sh",
   },
   {
-    name: "type=existing cluster, env=airgapped, phase=upgraded install, rbac=minimal rbac",
     backend_config: "existing-airgapped-upgrade-minimum-backend-config.tfvars",
     terraform_script: "existing-airgapped-upgrade-minimum.sh",
   }
@@ -96,31 +83,49 @@ for(const automationWorkspace of automationWorkspaces) {
         ... awsConfig,
         TF_WORKSPACE: automationWorkspace
       },
+      ignoreReturnCode: true,
       cwd: 'automation/jumpbox'
     });
-  const completionTime = new Date(completionTimestamp);
-  const currentTime = new Date();
-  // if(currentTime.getTime() - completionTime.getTime() > (1000 * 60 * 60 * 24)) {
-  if(currentTime.getTime() - completionTime.getTime() > (1)) {
-    for(const test of tests) {
-      await exec('terraform', [ 'init', '-backend-config', test.backend_config, '-reconfigure' ], {
-        env: awsConfig,
-        cwd: 'automation/cluster'
-      });
-      await exec(test.terraform_script, [ 'destroy' ], {
-        cwd: 'automation/cluster',
+
+  if(completionTimestamp) {
+    const currentTime = new Date();
+    // if(currentTime.getTime() - completionTime.getTime() > (1000 * 60 * 60 * 24)) {
+    if(currentTime.getTime() - completionTime.getTime() > (1)) {
+      for (const test of tests) {
+        await exec('terraform', ['init', '-backend-config', test.backend_config, '-reconfigure'], {
+          cwd: 'automation/cluster',
+          env: awsConfig
+        });
+        await exec(test.terraform_script, ['destroy'], {
+          cwd: 'automation/cluster',
+          env: {
+            ...awsConfig,
+            TF_WORKSPACE: automationWorkspace,
+          },
+        });
+        await exec('terraform', ['workspace', 'delete', automationWorkspace], {
+          cwd: 'automation/cluster',
+          env: {
+            ... awsConfig,
+            TF_WORKSPACE: 'default'
+          },
+        })
+      }
+      await exec('terraform', ['destroy', '-auto-approve'], {
+        cwd: 'automation/jumpbox',
         env: {
-          ... awsConfig,
+          ...awsConfig,
           TF_WORKSPACE: automationWorkspace,
         },
       });
     }
-    await exec('terraform', [ 'destroy', '-auto-approve' ], {
-      cwd: 'automation/jumpbox',
+
+    await exec('terraform', ['workspace', 'delete', automationWorkspace], {
       env: {
         ... awsConfig,
-        TF_WORKSPACE: automationWorkspace,
+        TF_WORKSPACE: 'default'
       },
-    });
+      cwd: 'automation/jumpbox'
+    })
   }
 }
